fix userInfo argument ValueError
